requirements.txt
absl-py==2.3.1
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
apscheduler==3.11.0
asyncio-mqtt==0.16.2
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
fastapi==0.104.1
flatbuffers==25.2.10
fonttools==4.59.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
hyperframe==6.1.0
idna==3.10
jax==0.7.0
jaxlib==0.7.0
jiter==0.10.0
kiwisolver==1.4.8
matplotlib==3.10.5
mediapipe==0.10.14
ml_dtypes==0.5.3
numpy>=2.0,<2.3
openai==1.47.0
opencv-contrib-python==4.12.0.88
onnxruntime>=1.19.2
insightface==0.7.3
opt_einsum==3.4.0
packaging==25.0
paho-mqtt==2.1.0
Pillow==10.4.0
postgrest==0.17.2
protobuf==4.25.8
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==21.0.1
PyYAML==6.0.2
realtime==2.5.3
requests==2.31.0
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.27.0
storage3==0.8.2
supabase==2.8.0
supafunc==0.6.2
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websockets==15.0.1
